test/mochitest/test_steps_basics.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  function assert_visible(id) {
    let tf = c.querySelector("section." + id);
    ok(tf, "has section for", id);

    let computed = c.defaultView.getComputedStyle(tf);
    is(computed.opacity, "1", id + " is visible");
    is(computed.pointerEvents, "auto", id + " has pointer events enabled");
  }

  function assert_hidden(id) {
    let tt = c.querySelector("section." + id);
    ok(tt, "has section for", id);

    let computedTo = c.defaultView.getComputedStyle(tt);
    is(computedTo.opacity, "0.2", id + " is not visible");
    is(computedTo.pointerEvents, "none", id + " has pointer events disabled");
  }

  function test_step_init() {
    return new Promise((resolve, reject) => {
      assert_visible("connect");
      assert_hidden("select");
      assert_hidden("flash");
      w.currentStep("connect");
      resolve();
    });
  }

  function test_step_connect() {
    return new Promise((resolve, reject) => {
      assert_visible("connect");
      assert_hidden("select");
      assert_hidden("flash");
      w.currentStep("select");
      resolve();
    });
  }

  function test_step_select() {
    return new Promise((resolve, reject) => {
      assert_visible("connect");
      assert_visible("select");
      assert_hidden("flash");
      w.currentStep("flash");
      resolve();
    });
  }

  function test_step_flash() {
    return new Promise((resolve, reject) => {
      assert_visible("connect");
      assert_visible("select");
      assert_visible("flash");
      resolve();
    });
  }

  function runTest() {
    populateAboutB2GInstaller()
      .then(() => test_step_init())
      .then(() => test_step_connect())
      .then(() => test_step_select())
      .then(() => test_step_flash())
      .then(() => SimpleTest.finish())
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
